requirements.txt
boto3
awscli
langchain
streamlit
faiss-cpu
pypdf
langchain_community
pandas
matplotlib
langchain-experimental
python-dotenv
plotly
botocore==1.31.68
python-dotenv==1.0.0
PyYAML==6.0.1
sentence-transformers==2.2.2
chromadb==0.4.14